+++ R/filter_ma.R
@@ -259,18 +259,18 @@ filter_ma <- function(
 
   # Folders---------------------------------------------------------------------
   path.folder <- generate_folder(
-    f = path.folder,
     rad.folder = "filter_ma",
+    path.folder = path.folder,
     internal = internal,
     file.date = file.date,
     verbose = verbose)
 
   # write the dots file
-  write_rad(
+  write_radiator_tsv(
     data = rad.dots,
-    path = path.folder,
-    filename = stringi::stri_join("radiator_filter_ma_args_", file.date, ".tsv"),
-    tsv = TRUE,
+    path.folder = path.folder,
+    filename = "radiator_filter_ma_args",
+    date = TRUE,
     internal = internal,
     write.message = "Function call and arguments stored in: ",
     verbose = verbose
@@ -438,7 +438,7 @@ filter_ma <- function(
       choices = c("MAC_GLOBAL_CORR", "MAF_GLOBAL_COUNT_CORR", "MAF_GLOBAL_DEPTH_CORR", "MAD_GLOBAL_CORR"),
       several.ok = FALSE
     )
-  if (rlang::has_name(x, stats)) nrow(dplyr::filter(.data = x, x[stats] >= threshold))
+  if (rlang::has_name(x, stats)) nrow(dplyr::filter(.data = x, x[[stats]] >= threshold))
   }#End how_many_markers
 
   end.seq <- ceiling(0.2 * n.diplo.samples)
@@ -453,7 +453,8 @@ filter_ma <- function(
     ) %>%
     readr::write_tsv(
       x = .,
-      file = file.path(path.folder, "mac.helper.table.tsv"))
+      file = file.path(path.folder, "mac.helper.table.tsv")
+      )
 
   # if (verbose) message("File written: mac.helper.table.tsv")
 
@@ -627,28 +628,34 @@ filter_ma <- function(
   #blacklist
 
   if (nrow(bl) > 0L) {
-    write_rad(
+    write_radiator_tsv(
       data = bl,
-      path = path.folder,
-      filename = stringi::stri_join("blacklist.markers.ma_", file.date, ".tsv"),
-      tsv = TRUE, internal = internal, verbose = verbose
+      path.folder = path.folder,
+      filename = "blacklist.markers.ma",
+      date = TRUE,
+      internal = internal,
+      write.message = "standard",
+      verbose = verbose
     )
   }
 
   # whitelist
   if (nrow(wl) > 0L) {
-    write_rad(
+    write_radiator_tsv(
       data = wl,
-      path = path.folder,
-      filename = stringi::stri_join("whitelist.markers.ma_", file.date, ".tsv"),
-      tsv = TRUE, internal = internal, verbose = verbose)
+      path.folder = path.folder,
+      filename = "whitelist.markers.ma",
+      date = TRUE,
+      internal = internal,
+      write.message = "standard",
+      verbose = verbose
+    )
   }
 
   if (!is.null(filename)) {
     if (data.type == "tbl_df") {
-      tidy.name <- stringi::stri_join(filename, ".rad")
-      if (verbose) message("Writing the MA filtered tidy data set: ", tidy.name)
-      write_rad(data = data, path = file.path(path.folder, tidy.name))
+      tidy.name <- stringi::stri_join(filename, ".arrow.parquet")
+      write_rad(data = data, filename = file.path(path.folder, tidy.name), verbose = verbose)
     }
   }
 
@@ -918,11 +925,12 @@ minor_allele_stats <- function(
             gds = data,
             individuals = FALSE,
             markers = TRUE,
-            dp = FALSE,
-            ad = TRUE,
+            coverage = FALSE,
+            allele.coverage = TRUE,
             coverage.stats = "sum",
             subsample.info = 1,
-            verbose = FALSE
+            verbose = FALSE,
+            parallel.core = parallel.core
           ) %$%
             m.info
         } else {
@@ -951,13 +959,14 @@ minor_allele_stats <- function(
         dplyr::mutate(N_ALLELES = purrr::map_int(.x = ac, .f = length)) %>%
         dplyr::filter(N_ALLELES > 2) %>%
         dplyr::select(dplyr::any_of(wanted.info)) %>%
-        write_rad(
+        write_radiator_tsv(
           data = .,
-          path = path.folder,
-          filename = "markers_number_alleles_problem.tsv",
-          tsv = TRUE,
+          path.folder = path.folder,
+          filename = "markers_number_alleles_problem",
+          date = TRUE,
+          internal = FALSE,
           write.message = "standard",
-          verbose = TRUE
+          verbose = verbose
         )
     }
     n.al.max <- NULL
@@ -1103,4 +1112,3 @@ ma_one <- carrier::crate(function(x) {
   mac.data$MAC_GLOBAL %<>% as.integer(.)
   return(mac.data)
 })#End ma_one
-