+++ R/genlight.R
@@ -153,7 +153,7 @@ tidy_genlight <- function(
 
   if (tidy) {
     if (write) {
-      filename.temp <- generate_filename(extension = "rad")
+      filename.temp <- generate_filename(extension = "arrow.parquet")
       filename.short <- filename.temp$filename.short
       filename.genlight <- filename.temp$filename
     }
@@ -186,25 +186,36 @@ tidy_genlight <- function(
       ) %$%
       input
 
-    if (write) radiator::write_rad(data = tidy.data, path = filename.genlight, verbose = verbose)
+    if (write) radiator::write_rad(data = tidy.data, filename = filename.genlight, verbose = verbose)
 
   }#End tidy genlight
 
   if (gds) {
     # generate the GDS --------------------------------------------------------------
     # markers %<>% dplyr::mutate(VARIANT_ID = as.integer(factor(MARKERS)))
+
+    # generate genotypes format for easy reading into GDS
+    tidy.data %<>%
+      dplyr::mutate(
+        GDS_A1 = dplyr::case_when(
+          GT_BIN == 0 ~ 0,
+          GT_BIN == 1 ~ 0,
+          GT_BIN == 2 ~ 1,
+          is.na(GT_BIN) ~ NA_integer_),
+        GDS_A2 = dplyr::case_when(
+          GT_BIN == 0 ~ 0,
+          GT_BIN == 1 ~ 1,
+          GT_BIN == 2 ~ 1,
+          is.na(GT_BIN) ~ NA_integer_)
+      )
+
     gds.filename <- radiator_gds(
       data.source = "genlight",
-      geno.coding = "alt.dos",
       genotypes = gt2array(
-        gt.bin = tidy.data$GT_BIN,
+        genotypes = tidy.data,
         n.ind = n.ind,
         n.snp = n.markers
       ),
-      # genotypes = tibble::as_tibble(data.frame(data) %>% t) %>%
-      #   tibble::add_column(.data = ., MARKERS = markers$MARKERS, .before = 1) %>%
-      #   dplyr::arrange(MARKERS) %>%
-      #   tibble::column_to_rownames(.data = ., var = "MARKERS"),
       strata = strata,
       biallelic = TRUE,
       markers.meta = markers,
@@ -224,16 +235,17 @@ tidy_genlight <- function(
 
 # write_genlight ----------------------------------------------------------------
 #' @name write_genlight
-#' @title Write a genlight object from a tidy data frame or GDS file or object.
-#' @description Write a genlight object from a tidy data frame or GDS file or object.
+#' @title Write a \code{genlight} object from: a tidy data frame, GDS file or object.
+#' @description Write a \code{genlight} object from a tidy data frame or GDS file or object.
 #' Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
-#' and might be of interest for users.
+#' and might be of interest for users. \code{genlight} is a formal (S4) class
+#' for storing genotypes of binary SNPs in a compact way, using a bit-level
+#' coding scheme. This storage is most efficient with haploid data,
+#' where the memory taken to represent data can be reduced more than 50 times.
+#' However, \code{genlight} can be used for any level of ploidy,
+#' and still remain an efficient storage mode.
 
 #' @inheritParams radiator_common_arguments
-
-#' @param biallelic (logical, optional) If you already know that the data is
-#' biallelic use this argument to speed up the function.
-#' Default: \code{biallelic = TRUE}.
 
 #' @param write (logical, optional) To write in the working directory the genlight
 #' object. The file is written with \code{radiator_genlight_DATE@TIME.RData} and
@@ -241,12 +253,17 @@ tidy_genlight <- function(
 #' Default: \code{write = FALSE}.
 
 #' @param dartr (logical, optional) For non-dartR users who wants to have a genlight
-#' object ready for the package. This option transfer or generates:
+#' object ready for the dartR package. This option transfer or generates:
 #' \code{CALL_RATE, AVG_COUNT_REF, AVG_COUNT_SNP, REP_AVG,
 #' ONE_RATIO_REF, ONE_RATIO_SNP}. These markers metadata are stored into
 #' the genlight slot: \code{genlight.obj@other$loc.metrics}.
 #' \strong{Use the radiator generated GDS data for best result}.
 #' Default: \code{dartr = FALSE}.
+
+
+#' @param biallelic (logical, optional) If you already know that the data is
+#' biallelic use this argument to speed up the function.
+#' Default: \code{biallelic = TRUE}.
 
 #' @export
 #' @rdname write_genlight
@@ -279,11 +296,11 @@ tidy_genlight <- function(
 
 write_genlight <- function(
     data,
-    biallelic = TRUE,
     write = FALSE,
     dartr = FALSE,
     verbose = FALSE,
-    parallel.core = parallel::detectCores() - 2
+    parallel.core = parallel::detectCores() - 2,
+    biallelic = TRUE
 ) {
 
 
@@ -316,65 +333,77 @@ write_genlight <- function(
 
     data.source <- radiator::extract_data_source(data)
     # markers.meta <- extract_markers_metadata(gds = data, whitelist = TRUE)
-    data.bk <- data
+    # data.bk <- data
     # data.bk -> data
-    if (dartr && !any(unique(c("1row", "2rows") %in% data.source))) {
+
+
+    # no longer needed
+    # if (dartr && !any(unique(c("1row", "2rows") %in% data.source))) {
       # counts data and data with read depth alleles depth info...
-      if (!"counts" %in% data.source) {
-        genotypes.meta <- radiator::extract_genotypes_metadata(gds = data, whitelist = TRUE)
-      } else {
-        genotypes.meta <- NULL
-      }
-      data <- gds2tidy(gds = data.bk,
-                       wide = FALSE,
-                       parallel.core = parallel.core,
-                       calibrate.alleles = FALSE)
+
+      # if (!"counts" %in% data.source) {
+      #   genotypes.meta <- radiator::extract_genotypes_metadata(gds = data, whitelist = TRUE)
+      # } else {
+      #   genotypes.meta <- NULL
+      # }
+
+      data <- gds2tidy(
+        gds = data,
+        wide = FALSE,
+        pop.id = FALSE,
+        parallel.core = parallel.core,
+        calibrate.alleles = FALSE
+      )
       markers.levels <- unique(data$MARKERS)
       ind.levels <- unique(data$INDIVIDUALS)
 
-      if (!"counts" %in% data.source) {
-        # genotypes.meta <- extract_coverage(# Note to my self... need to adapt the code here for the new extract function when you have the time
-        #   gds = data.bk,
-        #   individuals = FALSE,
-        #   depth.tibble = TRUE,
-        #   parallel.core = parallel.core
-        # )
-        # suppressWarnings(
-        #   data %<>%
-        #     dplyr::left_join(genotypes.meta, by = c("INDIVIDUALS", "MARKERS"))
-        # )
-      }
-
-      genotypes.meta <- NULL
-      want <- c("MARKERS", "CHROM", "LOCUS", "POS",
-                "REF", "ALT",
-                "CALL_RATE", "AVG_COUNT_REF", "AVG_COUNT_SNP", "REP_AVG",
-                "ONE_RATIO_REF", "ONE_RATIO_SNP")
+      # genotypes.meta <- NULL
+      want <- c("MARKERS", "CHROM", "LOCUS", "POS", "REF", "ALT",
+                "CALL_RATE",
+                "AVG_COUNT_REF",
+                "AVG_COUNT_SNP",
+                "REP_AVG",
+                "ONE_RATIO_REF",
+                "ONE_RATIO_SNP"
+                )
       markers.meta <- data %>%
         dplyr::select(tidyselect::any_of(want)) %>%
         dplyr::distinct(MARKERS, .keep_all = TRUE)
-    } else {
-      data <- radiator::extract_individuals_metadata(
-        gds = data, ind.field.select = c("INDIVIDUALS", "STRATA"), whitelist = TRUE) %>%
-        dplyr::bind_cols(
-          SeqArray::seqGetData(
-            gdsfile = data, var.name = "$dosage_alt") %>%
-            magrittr::set_colnames(x = ., value = markers.meta$MARKERS) %>%
-            tibble::as_tibble(.)
-        ) %>%
-        dplyr::rename(POP_ID = STRATA)
-    }
+
+      data %<>%
+        tidyr::pivot_wider(
+          id_cols = c(INDIVIDUALS, STRATA),
+          names_from = MARKERS,
+          values_from = GT_BIN
+        )
+    # }
+
+    # no longer needed
+    # else {
+    #   markers.meta <- radiator::extract_markers_metadata(
+    #     gds = data
+    #   )
+    #   data.check <- radiator::extract_individuals_metadata(
+    #     gds = data, ind.field.select = c("INDIVIDUALS", "STRATA"), whitelist = TRUE) %>%
+    #     dplyr::bind_cols(
+    #       SeqArray::seqGetData(
+    #         gdsfile = data, var.name = "$dosage_alt") %>%
+    #         magrittr::set_colnames(x = ., value = markers.meta$MARKERS) %>%
+    #         tibble::as_tibble(.)
+    #     ) %>%
+    #     dplyr::rename(POP_ID = STRATA)
+    # }
 
     # dartR-----------------------------------------------------------------------
     if (dartr) {
-      # That bits of code below generate what's nenessary for dartR
+      # That bits of code below generate what's necessary for dartR
       if (verbose) message("Calculating read depth for each SNP\n")
       if ("dart" %in% data.source) {
         # if (any(unique(c("1row", "2rows") %in% data.source))) {
         markers.meta %<>%
           dplyr::mutate(
             N_IND = SeqArray::seqApply(
-              gdsfile = data.bk,
+              gdsfile = data,
               var.name = "$dosage_alt",
               FUN = function(g) length(g[!is.na(g)]),
               margin = "by.variant", as.is = "integer",
@@ -385,7 +414,7 @@ write_genlight <- function(
             )
           ) %>%
           dplyr::ungroup(.)
-        data.bk <- NULL
+        # data.bk <- NULL
       } else {
         if (rlang::has_name(data, "READ_DEPTH")) {
           # dart 2 rows counts...
@@ -465,6 +494,7 @@ write_genlight <- function(
     # Detect if biallelic data ---------------------------------------------------
     if (is.null(biallelic)) biallelic <- radiator::detect_biallelic_markers(data)
     if (!biallelic) rlang::abort("genlight object requires biallelic genotypes")
+
     want <- c("MARKERS", "CHROM", "LOCUS", "POS", "REF", "ALT",
               "CALL_RATE", "AVG_COUNT_REF", "AVG_COUNT_SNP", "REP_AVG",
               "ONE_RATIO_REF", "ONE_RATIO_SNP")